+++ zcash_primitives/src/sapling.rs
@@ -2,6 +2,7 @@
 
 mod address;
 pub mod builder;
+pub mod bundle;
 pub mod circuit;
 pub mod constants;
 pub mod group_hash;
@@ -25,6 +26,7 @@ use constants::SPENDING_KEY_GENERATOR;
 use self::redjubjub::{PrivateKey, PublicKey, Signature};
 
 pub use address::PaymentAddress;
+pub use bundle::Bundle;
 pub use keys::{Diversifier, NullifierDerivingKey, ProofGenerationKey, SaplingIvk, ViewingKey};
 pub use note::{nullifier::Nullifier, Note, Rseed};
 pub use tree::{
