+++ zcash_primitives/src/transaction/components.rs
@@ -6,12 +6,13 @@ pub mod sapling;
 pub mod sprout;
 pub mod transparent;
 pub mod tze;
+
 pub use self::{
     amount::Amount,
-    sapling::{OutputDescription, SpendDescription},
     sprout::JsDescription,
     transparent::{OutPoint, TxIn, TxOut},
 };
+pub use crate::sapling::bundle::{OutputDescription, SpendDescription};
 
 #[cfg(feature = "zfuture")]
 pub use self::tze::{TzeIn, TzeOut};
