+++ zcash_proofs/src/prover.rs
@@ -6,11 +6,12 @@ use std::path::Path;
 use zcash_primitives::{
     sapling::{
         self,
+        bundle::GrothProofBytes,
         prover::{OutputProver, SpendProver},
         value::{NoteValue, ValueCommitTrapdoor},
         Diversifier, MerklePath, PaymentAddress, ProofGenerationKey, Rseed,
     },
-    transaction::components::{sapling::GrothProofBytes, GROTH_PROOF_SIZE},
+    transaction::components::GROTH_PROOF_SIZE,
 };
 
 use crate::{load_parameters, parse_parameters, OutputParameters, SpendParameters};
